+++ changelog.md
@@ -1,3 +1,18 @@
+### v2.12.0 (patch 8.2)
+---
+* Added patch 8.2 and Azshara's Eternal Palace as history mass delete options.
+
+
+* **Autopass**  
+* Added autopass data for trinkets in Crucible of Storms (a bit late, I know).
+* Added autopass data for trinkets in Eternal Palace.
+
+
+###### Bugfixes  
+* *Added a compatibility fix very old loot histories (#388).*
+* *Removed any Breath of Bwonsamdi that have been recorded to the history during alpha releases.*
+
+
 ### v2.11.0
 ---
 * **Loot Status**